+++ src/repo/repo.service.ts
@@ -2,13 +2,13 @@ import { Injectable, NotFoundException } from "@nestjs/common";
 import { ObjectLiteral, Repository, SelectQueryBuilder } from "typeorm";
 import { InjectRepository } from "@nestjs/typeorm";
 
-import { DbRepo } from "./entities/repo.entity";
 import { PageMetaDto } from "../common/dtos/page-meta.dto";
 import { PageDto } from "../common/dtos/page.dto";
-import { RepoOrderFieldsEnum, RepoPageOptionsDto } from "./dtos/repo-page-options.dto";
 import { OrderDirectionEnum } from "../common/constants/order-direction.constant";
 import { InsightFilterFieldsEnum } from "../insight/dtos/insight-options.dto";
 import { RepoFilterService } from "../common/filters/repo-filter.service";
+import { RepoOrderFieldsEnum, RepoPageOptionsDto } from "./dtos/repo-page-options.dto";
+import { DbRepo } from "./entities/repo.entity";
 import { RepoSearchOptionsDto } from "./dtos/repo-search-options.dto";
 
 @Injectable()