+++ src/user/user.entity.ts
@@ -10,16 +10,17 @@ import {
 } from "typeorm";
 import { ApiHideProperty } from "@nestjs/swagger";
 
+import {
+  ApiModelProperty,
+  ApiModelPropertyOptional,
+} from "@nestjs/swagger/dist/decorators/api-model-property.decorator";
 import { DbRepo } from "../repo/entities/repo.entity";
 import { DbRepoToUserVotes } from "../repo/entities/repo.to.user.votes.entity";
 import { DbRepoToUserStars } from "../repo/entities/repo.to.user.stars.entity";
 import { DbRepoToUserSubmissions } from "../repo/entities/repo.to.user.submissions.entity";
 import { DbRepoToUserStargazers } from "../repo/entities/repo.to.user.stargazers.entity";
-import {
-  ApiModelProperty,
-  ApiModelPropertyOptional,
-} from "@nestjs/swagger/dist/decorators/api-model-property.decorator";
 import { DbInsight } from "../insight/entities/insight.entity";
+import { DbUserNotification } from "./entities/user-notification.entity";
 import { DbUserHighlight } from "./entities/user-highlight.entity";
 import { DbUserHighlightReaction } from "./entities/user-highlight-reaction.entity";
 import { DbUserTopRepo } from "./entities/user-top-repo.entity";
@@ -542,4 +543,8 @@ export class DbUser extends BaseEntity {
   @ApiHideProperty()
   @OneToMany(() => DbUserTopRepo, (repoToUserTopRepos) => repoToUserTopRepos.user)
   public topRepos: DbUserTopRepo[];
+
+  @ApiHideProperty()
+  @OneToMany(() => DbUserNotification, (fromUserNotifications) => fromUserNotifications.from_user)
+  public from_user_notifications: DbUserNotification[];
 }